backend/requirements.txt
annotated-types==0.6.0
anyio==3.7.1
apt-xapian-index==0.49
async-generator==1.10
attrs==21.2.0
Babel==2.8.0
bcrypt==4.1.2
beautifulsoup4==4.10.0
black==21.10b0
blinker==1.4
certifi==2020.6.20
chardet==4.0.0
charset-normalizer==2.0.6
click==8.0.3
colorama==0.4.4
command-not-found==0.3
commonmark==0.9.1
cov-core==1.15.0
coverage==6.2
crypto==1.4.1
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
distro==1.7.0
distro-info==1.1+ubuntu0.1
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
docopt-ng==0.8.1
ecdsa==0.18.0
evdev==1.4.0
exceptiongroup==1.2.0
execnet==1.9.0
fastapi==0.109.0
fastapi-login==1.9.3
flake8==4.0.1
fuse-python==1.0.2
gdtoolkit==4.1.0
gpg==1.16.0
gyp==0.1
h11==0.13.0
html5lib==1.1
httpcore==0.14.5
httplib2==0.20.2
httptools==0.6.1
httpx==0.22.0
idna==3.3
img2pdf==0.4.2
importlib-metadata==4.6.4
iniconfig==1.1.1
isort==5.6.4
jeepney==0.7.1
Jinja2==3.0.3
jsonschema==3.2.0
kazam==1.4.5
keyring==23.5.0
language-selector==0.1
lark==1.1.5
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libevdev==0.5
livereload==2.6.3
lxml==4.8.0
mando==0.7.1
Markdown==3.3.6
markdown-it-py==3.0.0
MarkupSafe==2.0.1
mccabe==0.6.1
mdurl==0.1.2
meson==0.61.2
mkdocs==1.1.2
more-itertools==8.10.0
mypy==0.942
mypy-extensions==0.4.3
Naked==0.1.31
nala==0.11.1
netifaces==0.11.0
nodeenv==1.8.0
nose2==0.9.2
numpy==1.21.5
oauthlib==3.2.0
olefile==0.46
outcome==1.3.0.post0
packaging==21.3
pathspec==0.9.0
pdfarranger==1.8.2
pexpect==4.8.0
pikepdf==5.0.1+dfsg
Pillow==9.0.1
platformdirs==2.5.1
pluggy==0.13.0
prisma==0.12.0
psutil==5.9.0
ptyprocess==0.7.0
py==1.10.0
pyasn1==0.5.1
pycairo==1.20.1
pycodestyle==2.8.0
pycryptodome==3.15.0
pycups==2.0.1
pydantic==2.5.3
pydantic_core==2.14.6
pyflakes==2.4.0
Pygments==2.17.2
PyGObject==3.42.1
pyinotify==0.9.6
PyJWT==2.8.0
pylibacl==0.6.0
pyparsing==2.4.7
PyQt5==5.15.6
PyQt5-sip==12.9.1
pyrsistent==0.18.1
pytest==6.2.5
pytest-cov==3.0.0
pytest-forked==1.4.0
pytest-sugar==0.9.4
pytest-xdist==2.5.0
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
python-dotenv==0.19.2
python-jose==3.3.0
python-magic==0.4.24
python-multipart==0.0.6
pytz==2022.1
pyudev==0.22.0
pyxattr==0.7.2
pyxdg==0.27
PyYAML==5.4.1
radon==6.0.1
regex==2023.8.8
reportlab==3.6.8
requests==2.25.1
rfc3986==1.5.0
rich==13.7.0
rsa==4.9
scour==0.38.2
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
setproctitle==1.2.2
shellescape==3.8.1
shellingham==1.4.0
six==1.16.0
sniffio==1.2.0
socksio==1.0.0
sortedcontainers==2.4.0
soupsieve==2.3.1
ssh-import-id==5.11
starlette==0.35.1
StrEnum==0.4.15
systemd-python==234
termcolor==2.3.0
texttable==1.6.4
toml==0.10.2
tomli==1.2.2
tomlkit==0.12.3
tornado==6.1
tqdm==4.64.1
trio==0.21.0
typed-ast==1.4.3
typer==0.4.0
typing_extensions==4.9.0
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
uvicorn==0.27.0
uvloop==0.19.0
wadllib==1.3.6
watchfiles==0.21.0
webencodings==0.5.1
websocket-client==1.2.3
websockets==12.0
WoeUSB-ng @ file:///home/matthew/temp/WoeUSB-ng
wxPython==4.0.7
xdg==5
xkit==0.0.0
zipp==1.0.0
